+++ core/server/api_container/server/service_network/service_network.go
@@ -79,4 +79,6 @@ type ServiceNetwork interface {
 	GetIPAddressForService(serviceID service.ServiceID) (net.IP, bool)
 
 	RenderTemplates(templatesAndDataByDestinationRelFilepath map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData) (enclave_data_directory.FilesArtifactUUID, error)
+
+	UploadFilesArtifact(data []byte) (enclave_data_directory.FilesArtifactUUID, error)
 }