+++ internal_testsuites/typescript/src/test_helpers/test_helpers.ts
@@ -19,6 +19,7 @@ import * as fs from 'fs';
 import * as path from "path";
 import * as os from "os";
 import axios from "axios";
+import {GetArgs, GetResponse, UpsertArgs} from "example-datastore-server-api-lib";
 
 const CONFIG_FILENAME = "config.json"
 const CONFIG_MOUNTPATH_ON_API_CONTAINER = "/config"
@@ -58,6 +59,31 @@ const WAIT_FOR_AVAILABILITY_BODY_TEXT     = ""
 const USER_SERVICE_MOUNT_POINT_FOR_TEST_FILES_ARTIFACT  = "/static"
 
 const FILE_SERVER_PORT_SPEC = new PortSpec( FILE_SERVER_PRIVATE_PORT_NUM, PortProtocol.TCP )
+
+// for validating data store is healthy
+/*
+NOTE: on 2022-05-16 the Go version failed with the following error so we bumped the num polls to 20.
+
+time="2022-05-16T23:58:21Z" level=info msg="Sanity-checking that all 4 datastore services added via the module work as expected..."
+--- FAIL: TestModule (21.46s)
+    module_test.go:81:
+            Error Trace:	module_test.go:81
+            Error:      	Received unexpected error:
+                            The service didn't return a success code, even after 15 retries with 1000 milliseconds in between retries
+                             --- at /home/circleci/project/internal_testsuites/golang/test_helpers/test_helpers.go:179 (WaitForHealthy) ---
+                            Caused by: rpc error: code = Unavailable desc = connection error: desc = "transport: Error while dialing dial tcp 127.0.0.1:49188: connect: connection refused"
+            Test:       	TestModule
+            Messages:   	An error occurred waiting for the datastore service to become available
+
+NOTE: On 2022-05-21 the Go version failed again at 20s. I opened the enclave logs and it's weird because nothing is failing and
+the datastore service is showing itself as up *before* we even start the check-if-available wait. We're in crunch mode
+so I'm going to bump this up to 30s, but I suspect there's some sort of nondeterministic underlying failure happening.
+ */
+const WAIT_FOR_STARTUP_MAX_POLLS = 30
+const MILLIS_BETWEEN_AVAILABILITY_RETRIES = 1000
+const TEST_DATASTORE_KEY = "my-key"
+const TEST_DATASTORE_VALUE = "test-value"
+
 
 class StartFileServerResponse  {
     fileServerPublicIp: string
@@ -372,4 +398,90 @@ async function getFileContents(ipAddress: string, portNum: number, relativeFilep
 
     const bodyStr = String(response.data)
     return ok(bodyStr)
+}
+
+export async function validateDataStoreServiceIsHealthy(enclaveContext : EnclaveContext, serviceId: string, portId: string): Promise<Result<null, Error>> {
+    const getServiceContextResult = await enclaveContext.getServiceContext(serviceId)
+    if (getServiceContextResult.isErr()) {
+        log.error(`An error occurred getting the service context for service '${serviceId}'; this indicates that the module says it created a service that it actually didn't`)
+        throw getServiceContextResult.error
+    }
+    const serviceContext = getServiceContextResult.value
+    const ipAddr = serviceContext.getMaybePublicIPAddress()
+    const publicPort: undefined | PortSpec = serviceContext.getPublicPorts().get(portId)
+
+    if (publicPort === undefined) {
+        throw new Error(`Expected to find public port '${portId}' on datastore service '${serviceId}', but none was found`)
+    }
+
+    const {
+        client: datastoreClient,
+        clientCloseFunction: datastoreClientCloseFunction
+    } = createDatastoreClient(ipAddr, publicPort.number);
+
+    try {
+        const waitForHealthyResult = await waitForHealthy(datastoreClient, WAIT_FOR_STARTUP_MAX_POLLS, MILLIS_BETWEEN_AVAILABILITY_RETRIES);
+        if (waitForHealthyResult.isErr()) {
+            log.error(`An error occurred waiting for the datastore service '${serviceId}' to become available`);
+            throw waitForHealthyResult.error
+        }
+
+        const upsertArgs = new UpsertArgs();
+        upsertArgs.setKey(TEST_DATASTORE_KEY)
+        upsertArgs.setValue(TEST_DATASTORE_VALUE)
+
+        const upsertResponseResultPromise:Promise<Result<google_protobuf_empty_pb.Empty, Error>> = new Promise((resolve, _unusedReject) => {
+            datastoreClient.upsert(upsertArgs, (error: grpc.ServiceError | null, response?: google_protobuf_empty_pb.Empty) => {
+                if (error === null) {
+                    if (!response) {
+                        resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")));
+                    } else {
+                        resolve(ok(response));
+                    }
+                } else {
+                    resolve(err(error));
+                }
+            })
+        })
+
+        const upsertResponseResult = await upsertResponseResultPromise;
+        if(upsertResponseResult.isErr()){
+            log.error(`An error occurred adding the test key to datastore service "${serviceId}"`)
+            throw upsertResponseResult.error
+        }
+
+        const getArgs = new GetArgs();
+        getArgs.setKey(TEST_DATASTORE_KEY)
+
+        const getResponseResultPromise:Promise<Result<GetResponse, Error>> = new Promise((resolve, _unusedReject) => {
+            datastoreClient.get(getArgs, (error: grpc.ServiceError | null, response?: GetResponse) => {
+                if (error === null) {
+                    if (!response) {
+                        resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")));
+                    } else {
+                        resolve(ok(response));
+                    }
+                } else {
+                    resolve(err(error));
+                }
+            })
+        })
+
+        const getResponseResult = await getResponseResultPromise;
+        if(getResponseResult.isErr()){
+            log.error(`An error occurred getting the test key to datastore service "${serviceId}"`)
+            throw getResponseResult.error
+        }
+
+        const getResponse = getResponseResult.value;
+        const actualValue = getResponse.getValue();
+        if(actualValue !== TEST_DATASTORE_VALUE){
+            log.error(`Datastore service "${serviceId}" is storing value "${actualValue}" for the test key, which doesn't match the expected value ""${TEST_DATASTORE_VALUE}`)
+        }
+
+    } finally {
+        datastoreClientCloseFunction()
+    }
+
+    return ok(null)
 }